+++ org.palladiosimulator.experimentautomation/src/org/palladiosimulator/experimentautomation/experiments/Variation.java
@@ -101,12 +101,12 @@ public interface Variation extends EObject {
      * <!-- end-user-doc -->
      *
      * @return the value of the '<em>Min Value</em>' attribute.
-     * @see #setMinValue(long)
+     * @see #setMinValue(double)
      * @see org.palladiosimulator.experimentautomation.experiments.ExperimentsPackage#getVariation_MinValue()
      * @model required="true" ordered="false"
      * @generated
      */
-    long getMinValue();
+    double getMinValue();
 
     /**
      * Sets the value of the '
@@ -118,7 +118,7 @@ public interface Variation extends EObject {
      * @see #getMinValue()
      * @generated
      */
-    void setMinValue(long value);
+    void setMinValue(double value);
 
     /**
      * Returns the value of the '<em><b>Max Value</b></em>' attribute. <!-- begin-user-doc -->
@@ -129,12 +129,12 @@ public interface Variation extends EObject {
      * <!-- end-user-doc -->
      *
      * @return the value of the '<em>Max Value</em>' attribute.
-     * @see #setMaxValue(long)
+     * @see #setMaxValue(double)
      * @see org.palladiosimulator.experimentautomation.experiments.ExperimentsPackage#getVariation_MaxValue()
      * @model required="true" ordered="false"
      * @generated
      */
-    long getMaxValue();
+    double getMaxValue();
 
     /**
      * Sets the value of the '
@@ -146,7 +146,7 @@ public interface Variation extends EObject {
      * @see #getMaxValue()
      * @generated
      */
-    void setMaxValue(long value);
+    void setMaxValue(double value);
 
     /**
      * Returns the value of the '<em><b>Max Variations</b></em>' attribute. <!-- begin-user-doc -->
