+++ org.palladiosimulator.experimentautomation/src/org/palladiosimulator/experimentautomation/experiments/impl/NestedIntervalsValueProviderImpl.java
@@ -34,7 +34,7 @@ public class NestedIntervalsValueProviderImpl extends ValueProviderImpl implemen
      * @generated
      * @ordered
      */
-    protected static final long MIN_VALUE_EDEFAULT = 0L;
+    protected static final double MIN_VALUE_EDEFAULT = 0.0;
 
     /**
      * The cached value of the '{@link #getMinValue() <em>Min Value</em>}' attribute. <!--
@@ -44,7 +44,7 @@ public class NestedIntervalsValueProviderImpl extends ValueProviderImpl implemen
      * @generated
      * @ordered
      */
-    protected long minValue = MIN_VALUE_EDEFAULT;
+    protected double minValue = MIN_VALUE_EDEFAULT;
 
     /**
      * The default value of the '{@link #getMaxValue() <em>Max Value</em>}' attribute. <!--
@@ -54,7 +54,7 @@ public class NestedIntervalsValueProviderImpl extends ValueProviderImpl implemen
      * @generated
      * @ordered
      */
-    protected static final long MAX_VALUE_EDEFAULT = 0L;
+    protected static final double MAX_VALUE_EDEFAULT = 0.0;
 
     /**
      * The cached value of the '{@link #getMaxValue() <em>Max Value</em>}' attribute. <!--
@@ -64,7 +64,7 @@ public class NestedIntervalsValueProviderImpl extends ValueProviderImpl implemen
      * @generated
      * @ordered
      */
-    protected long maxValue = MAX_VALUE_EDEFAULT;
+    protected double maxValue = MAX_VALUE_EDEFAULT;
 
     /**
      * <!-- begin-user-doc --> <!-- end-user-doc -->
@@ -91,7 +91,7 @@ public class NestedIntervalsValueProviderImpl extends ValueProviderImpl implemen
      * @generated
      */
     @Override
-    public long getMinValue() {
+    public double getMinValue() {
         return this.minValue;
     }
 
@@ -101,8 +101,8 @@ public class NestedIntervalsValueProviderImpl extends ValueProviderImpl implemen
      * @generated
      */
     @Override
-    public void setMinValue(final long newMinValue) {
-        final long oldMinValue = this.minValue;
+    public void setMinValue(final double newMinValue) {
+        final double oldMinValue = this.minValue;
         this.minValue = newMinValue;
         if (this.eNotificationRequired()) {
             this.eNotify(new ENotificationImpl(this, Notification.SET,
@@ -116,7 +116,7 @@ public class NestedIntervalsValueProviderImpl extends ValueProviderImpl implemen
      * @generated
      */
     @Override
-    public long getMaxValue() {
+    public double getMaxValue() {
         return this.maxValue;
     }
 
@@ -126,8 +126,8 @@ public class NestedIntervalsValueProviderImpl extends ValueProviderImpl implemen
      * @generated
      */
     @Override
-    public void setMaxValue(final long newMaxValue) {
-        final long oldMaxValue = this.maxValue;
+    public void setMaxValue(final double newMaxValue) {
+        final double oldMaxValue = this.maxValue;
         this.maxValue = newMaxValue;
         if (this.eNotificationRequired()) {
             this.eNotify(new ENotificationImpl(this, Notification.SET,
@@ -160,10 +160,10 @@ public class NestedIntervalsValueProviderImpl extends ValueProviderImpl implemen
     public void eSet(final int featureID, final Object newValue) {
         switch (featureID) {
         case ExperimentsPackage.NESTED_INTERVALS_VALUE_PROVIDER__MIN_VALUE:
-            this.setMinValue((Long) newValue);
+            this.setMinValue((Double) newValue);
             return;
         case ExperimentsPackage.NESTED_INTERVALS_VALUE_PROVIDER__MAX_VALUE:
-            this.setMaxValue((Long) newValue);
+            this.setMaxValue((Double) newValue);
             return;
         }
         super.eSet(featureID, newValue);
